Fix MSVC warnings

+++ example/histogram.cpp
@@ -30,7 +30,7 @@ flux::generator<int> randn(T mean, T stddev = 1.0)
     std::normal_distribution dist(mean, stddev);
 
     while (true) {
-        co_yield std::round(dist(rng));
+        co_yield static_cast<int>(std::round(dist(rng)));
     }
 }
